Move to insights_client_conf parser
